requirements.txt
streamlit
python-dotenv
nest_asyncio
google-generativeai
requests
pydantic
